% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_gitignore.R
\name{add_code_of_conduct}
\alias{add_code_of_conduct}
\title{Add a Code of Conduct to a directory.}
\usage{
add_code_of_conduct(path = getwd())
}
\arguments{
\item{path}{The path to the directory where the code of conduct file will be
created.}
}
\description{
Add a Code of Conduct to a directory.
}
