% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_gitignore.R
\name{add_gitignore}
\alias{add_gitignore}
\title{Add .gitignore to a directory.}
\usage{
add_gitignore(path = getwd())
}
\arguments{
\item{path}{The path to the directory where the git ignore file will be
created.}
}
\description{
Add .gitignore to a directory.
}
