% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_team_members.R
\name{add_team_members}
\alias{add_team_members}
\title{Add GitHub Users to a Team}
\usage{
add_team_members(team, members, org = "openscapes")
}
\arguments{
\item{team}{The name of the team.}

\item{members}{A vector of GitHub usernames.}

\item{org}{The GitHub organization that owns the team and the repository.}
}
\value{
list of responses from the GitHub API
}
\description{
This also adds them to the organization that contains the team. All users
will get an invitation by email and on GitHub.com, that they need to accept.
This invitation expires after seven days.
}
\examples{
\dontrun{
add_team_members("2021-ilm-rotj-team", members = c("not-ateucher", "seankross"))
}
}
