% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_team_to_repo.R
\name{add_team_to_repo}
\alias{add_team_to_repo}
\title{Add a GitHub Team to a Repository}
\usage{
add_team_to_repo(repository, team, org = "openscapes")
}
\arguments{
\item{repository}{The name of the GitHub repository.}

\item{team}{The name of the team.}

\item{org}{The GitHub organization that owns the team and the repository.}
}
\description{
Add a GitHub Team to a Repository
}
\examples{
\dontrun{

kyber::add_team_to_repo("2021-ilm-rotj", "2021-ilm-rotj-team")
}
}
