% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_duration.R
\name{check_duration}
\alias{check_duration}
\title{Check the duration of call}
\usage{
check_duration(
  registry_url,
  cohort_id,
  call_number,
  cohort_sheet = "cohort_metadata",
  call_sheet = "call_metadata"
)
}
\arguments{
\item{registry_url}{The URL to the cohort registry.}

\item{cohort_id}{The ID of the cohort.}

\item{call_number}{The call number of the agenda.}

\item{cohort_sheet}{The sheet in the registry with cohort information.}

\item{call_sheet}{The sheet in the registry with information about
individual calls.}
}
\description{
Check the duration of call
}
