% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_certificate.R
\name{create_batch_certificates}
\alias{create_batch_certificates}
\title{Render a batch of certificates for a Champions Cohort}
\usage{
create_batch_certificates(
  registry,
  participants,
  cohort_name,
  cohort_type = c("standard", "nmfs"),
  output_dir = "."
)
}
\arguments{
\item{registry}{\code{data.frame} from Google Sheet of Champions registry
(\code{"OpenscapesChampionsCohortRegistry"})}

\item{participants}{\code{data.frame} from Google Sheet of Champions participants
(\code{"OpenscapesParticipantsMainList"})}

\item{cohort_name}{Name of the cohort as it appears in \code{registry} and
\code{participants}}

\item{cohort_type}{What kind of cohort are the certificates for? This will
choose the appropriate certificate template: \code{"standard"} (default) or
\code{"nmfs"}.}

\item{output_dir}{output directory for certificates. Default \code{"."}}
}
\value{
path to the directory containing the certificates
}
\description{
Given a data.frame of Champions participants and a Cohort registry,
create a certificate for each participant
}
\examples{
\dontrun{
registry <- read_sheet("path-to-registry")
participants <- read_sheet("path-to-participants")

create_batch_certificates(
  registry,
  particpants,
  "2023-fred-hutch",
  "~/Desktop/fred-hutch-certificates"
)
}
}
