% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_certificate.R
\name{create_certificate}
\alias{create_certificate}
\title{Render certificates of completion for a participant in a cohort}
\usage{
create_certificate(
  cohort_name,
  first_name,
  last_name,
  start_date,
  end_date,
  cohort_website,
  cohort_type = c("standard", "nmfs"),
  output_dir = ".",
  quiet = TRUE,
  ...
)
}
\arguments{
\item{cohort_name}{The name of the cohort}

\item{first_name}{First name of participant}

\item{last_name}{Last name of participant}

\item{start_date}{cohort start date. \code{Date} or character in a standard format
(eg., YYYY-MM-DD)}

\item{end_date}{cohort end date. \code{Date} or character in a standard format
(eg., YYYY-MM-DD)}

\item{cohort_website}{cohort website}

\item{cohort_type}{What kind of cohort are the certificates for? This will
choose the appropriate certificate template: \code{"standard"} (default) or
\code{"nmfs"}.}

\item{output_dir}{output directory for certificates. Default \code{"."}}

\item{quiet}{Suppress quarto warnings and other messages. Default \code{TRUE}.
Set to \code{FALSE} to help debug if any errors occur.}

\item{...}{Other parameters passed on to \code{\link[quarto:quarto_render]{quarto::quarto_render()}}}
}
\value{
Saves the file to the specified directory, and returns the path to
the file
}
\description{
Render certificates of completion for a participant in a cohort
}
\examples{
\dontrun{
create_certificate(
  cohort_name = "2023-fred-hutch",
  first_name = "FirstName",
  last_name = "LastName",
  start_date = "2023-09-19",
  end_date = "2023-10-19",
  cohort_website = "https://openscapes.github.io/2023-fred-hutch/",
  cohort_type = "standard"
)
}
}
