% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_github_clinic.R
\name{create_github_clinic}
\alias{create_github_clinic}
\title{Create GitHub Clinic Files}
\usage{
create_github_clinic(names, path = getwd())
}
\arguments{
\item{names}{A vector of names for the markdown files that should be created.
The \code{.md} extension will be added automatically. The names should be unique.
If there exists a file with the same name in the target directory, the new file
will be prepended with \verb{_duplicate_} and a warning will be issued.}

\item{path}{Path to the directory where \code{github-clinic} should be created.}
}
\description{
Create GitHub Clinic Files
}
\examples{
\dontrun{

create_github_clinic(names = c("julia", "erin"))

file.exists("github-clinic")
#> TRUE

list.files("github-clinic")
#> "erin.md"  "julia.md"
}
}
