% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_readme.R
\name{create_readme}
\alias{create_readme}
\title{Create a README.Rmd from a Kyber Template}
\usage{
create_readme(
  file = "README.Rmd",
  template = "openscapes-cohort-readme",
  edit = TRUE
)
}
\arguments{
\item{file}{File name for the draft}

\item{template}{Template to use as the basis for the draft. This is either
the full path to a template directory or the name of a template directory
within the \code{rmarkdown/templates} directory of a package.}

\item{edit}{\code{TRUE} to edit the template immediately}
}
\description{
Create a README.Rmd from a Kyber Template
}
\details{
Valid values for the \code{template} argument include \code{"openscapes-cohort-readme"}.
}
\examples{
\dontrun{

kyber::create_readme()
}
}
