% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_team.R
\name{create_team}
\alias{create_team}
\title{Create a Team on GitHub for an Organization}
\usage{
create_team(name, maintainers, org = "openscapes", visible = TRUE)
}
\arguments{
\item{name}{The name of the team.}

\item{maintainers}{One or more GitHub usernames that will be the team maintainers.}

\item{org}{The GitHub organization that will own the team.}

\item{visible}{Should the team be visible to every member of the organization?}
}
\description{
Create a Team on GitHub for an Organization
}
\examples{
\dontrun{

# One maintainer
kyber::create_team("2021-ilm-rotj-team", maintainers = "jules32")

# Multiple maintainers
kyber::create_team("2021-ilm-rotj-team", maintainers = c("jules32", "seankross"))
}
}
