% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ky_fmt.R
\name{formatting}
\alias{ky_fmt_wiki_link}
\alias{ky_fmt_schedule}
\title{Readme Formatting}
\usage{
ky_fmt_wiki_link(link)

ky_fmt_schedule(schedule)
}
\arguments{
\item{link}{A url for an online wiki.}

\item{schedule}{A list of data frames that will be combined, row-wise.}
}
\description{
Functions to help format yaml data nicely in Readme templates.
}
