% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ky_add_repo_to_team.R
\name{ky_add_repo_to_team}
\alias{ky_add_repo_to_team}
\title{Add a GitHub Repository to a Team}
\usage{
ky_add_repo_to_team(repository, team, org = "openscapes")
}
\arguments{
\item{repository}{The name of the GitHub repository.}

\item{team}{The name of the team.}

\item{org}{The GitHub organization that owns the team and the repository.}
}
\description{
Add a GitHub Repository to a Team
}
\examples{
\dontrun{

kyber::ky_add_repo_to_team("2021-ilm-rotj", "2021-ilm-rotj-team")
}
}
