% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ky_add_team_members.R
\name{ky_add_team_members}
\alias{ky_add_team_members}
\title{Add GitHub Users to a Team}
\usage{
ky_add_team_members(team, members, org = "openscapes")
}
\arguments{
\item{team}{The name of the team.}

\item{members}{A vector of GitHub usernames.}

\item{org}{The GitHub organization that owns the team and the repository.}
}
\description{
Add GitHub Users to a Team
}
\examples{
\dontrun{

kyber::ky_add_team_members("2021-ilm-rotj-team", members = c("erinmr", "seankross"))
}
}
