% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ky_create_github_clinic.R
\name{ky_create_github_clinic}
\alias{ky_create_github_clinic}
\title{Create GitHub Clinic Files}
\usage{
ky_create_github_clinic(names, path = getwd())
}
\arguments{
\item{names}{A vector of names for the markdown files that should be created.
The \code{.md} extension will be added automatically.}

\item{path}{Path to the directory where \code{github-clinic} should be created.}
}
\description{
Create GitHub Clinic Files
}
\examples{
\dontrun{

ky_create_github_clinic(names = c("julia", "erin"))

file.exists("github-clinic")
#> TRUE

list.files("github-clinic")
#> "erin.md"  "julia.md"
}
}
