% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ky_create_repo.R
\name{ky_create_repo}
\alias{ky_create_repo}
\title{Create a GitHub repository with a README.Rmd}
\usage{
ky_create_repo(
  name,
  org = "openscapes",
  path = getwd(),
  template = "openscapes-cohort-readme",
  edit = TRUE
)
}
\arguments{
\item{name}{The name of the repository.}

\item{org}{The GitHub organization that will own the repository.}

\item{path}{The path to a directory where the repository will be created.}

\item{template}{Template to use as the basis for the draft. This is either
the full path to a template directory or the name of a template directory
within the \code{rmarkdown/templates} directory of a package.}

\item{edit}{\code{TRUE} to edit the template immediately}
}
\description{
Create a GitHub repository with a README.Rmd
}
\details{
Valid values for the \code{template} argument include \code{"openscapes-cohort-readme"},
or \code{FALSE} if you want to create a repository with no \code{README.md}.
}
\examples{
\dontrun{

kyber::ky_create_repo("2021-ilm-rotj")
}
}
