% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kyber-package.R
\docType{package}
\name{kyber-package}
\alias{kyber}
\alias{kyber-package}
\title{kyber: Set Up Learning Cohorts on GitHub}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides tools for setting up Openscapes learning cohorts, including creating agendas, creating and modifying GitHub org/team memberships, setting up cohort GitHub repositories, and creating completion certificates.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://openscapes.github.io/kyber/}
  \item \url{https://github.com/Openscapes/kyber}
  \item Report bugs at \url{https://github.com/Openscapes/kyber/issues}
}

}
\author{
\strong{Maintainer}: Sean Kross \email{sean@seankross.com} (\href{https://orcid.org/0000-0001-5215-0316}{ORCID})

Authors:
\itemize{
  \item Sean Kross \email{sean@seankross.com} (\href{https://orcid.org/0000-0001-5215-0316}{ORCID})
  \item Julia Lowndes \email{julia@openscapes.org} (\href{https://orcid.org/0000-0003-1682-3872}{ORCID})
  \item Andy Teucher \email{andy.teucher@gmail.com} (\href{https://orcid.org/0000-0002-7840-692X}{ORCID})
  \item Stefanie Butland \email{stefaniebutland@gmail.com} (\href{https://orcid.org/0000-0002-5427-8951}{ORCID})
}

Other contributors:
\itemize{
  \item Openscapes (\href{https://ror.org/022je8241}{ROR}) [copyright holder]
}

}
\keyword{internal}
