% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kyber_file.R
\name{kyber_file}
\alias{kyber_file}
\title{Get the paths to files installed with kyber}
\usage{
kyber_file(path = NULL)
}
\arguments{
\item{path}{The name of the file. If no argument is provided then
all of the example files will be listed.}
}
\value{
A vector of file paths
}
\description{
This function allows you to quickly access files that are installed with
kyber.
}
\examples{
kyber_file("ari_intro.Rmd")

}
