% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_teams.R
\name{list_team_members}
\alias{list_team_members}
\title{List members of a GitHub Team}
\usage{
list_team_members(
  team,
  org = "openscapes",
  names_only = TRUE,
  members = c("members", "invitations"),
  ...
)
}
\arguments{
\item{team}{The name of the team.}

\item{org}{The GitHub organization that owns the team and the repository.}

\item{names_only}{Should only the team member names be returned (as a
character vector; \code{TRUE}, the default), or should all of the team member
metadata be returned as a data.frame?}

\item{members}{Should current members (\code{"members"}) be returned, or pending
invitations (\code{"invitations"}) invitations be returned? Default \code{"members"}.}

\item{...}{passed on to \code{\link[gh:gh]{gh::gh()}}}
}
\value{
a character vector of team member GitHub usernames if \code{names_only = TRUE}, otherwise a \code{gh_response} object containing team member information
}
\description{
List members of a GitHub Team
}
\examples{
\dontrun{
  list_team_members(team = "2023-superdogs-test-cohort", org = "openscapes")
  list_team_members(team = "2023-superdogs-test-cohort", org = "openscapes",
                    names_only = FALSE)
  list_team_members(team = "2023-superdogs-test-cohort", org = "openscapes",
                    members = "invitations")
}
}
