% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_team_members.R
\name{list_teams}
\alias{list_teams}
\title{List teams in a GitHub organization}
\usage{
list_teams(org = "openscapes", names_only = TRUE, ...)
}
\arguments{
\item{org}{The GitHub organization that owns the team and the repository.}

\item{names_only}{Should only the team names be returned (as a character vector; \code{TRUE}, the default),
or should all of the team metadata be returned?}

\item{...}{passed on to \code{\link[gh:gh]{gh::gh()}}}
}
\value{
a character vector of team names if \code{names_only = TRUE}, otherwise
a \code{gh_response} object containing team information
}
\description{
List teams in a GitHub organization
}
\examples{
\dontrun{
  list_teams(org = "openscapes")
  list_teams(org = "openscapes", names_only = FALSE)
}
}
