% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_team_members.R
\name{remove_org_members}
\alias{remove_org_members}
\title{Remove GitHub organization members}
\usage{
remove_org_members(members, org = "openscapes")
}
\arguments{
\item{members}{A vector of GitHub usernames.}

\item{org}{The GitHub organization that owns the team and the repository.}
}
\value{
list of responses from the GitHub API
}
\description{
Remove GitHub organization members
}
\examples{
\dontrun{
 remove_org_members(members = "not-ateucher")
}
}
