% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_team_members.R
\name{remove_team_members}
\alias{remove_team_members}
\title{Remove GitHub Users from a Team}
\usage{
remove_team_members(team, members, org = "openscapes")
}
\arguments{
\item{team}{The name of the team.}

\item{members}{A vector of GitHub usernames.}

\item{org}{The GitHub organization that owns the team and the repository.}
}
\value{
list of responses from the GitHub API
}
\description{
Remove GitHub Users from a Team
}
\examples{
\dontrun{
remove_team_members("2021-ilm-rotj-team", members = "not-ateucher"))
}
}
