% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_certificate.R
\name{render_certificate}
\alias{render_certificate}
\title{Render certificates of completion for a participant in a cohort}
\usage{
render_certificate(
  cohort_name,
  participant_name,
  start_date,
  end_date,
  cohort_website,
  output_dir = "."
)
}
\arguments{
\item{cohort_name}{The name of the cohort}

\item{participant_name}{The name of the participant}

\item{start_date}{cohort start date}

\item{end_date}{cohort end date}

\item{cohort_website}{cohort website}

\item{output_dir}{output directory for certificates. Default \code{"."}}
}
\value{
Saves the file to your current working directory, and
returns the path to the file
}
\description{
Render certificates of completion for a participant in a cohort
}
\examples{

render_certificate(cohort_name = "2023-fred-hutch",
                   participant_name = "Name",
                   start_date = "Sep 19",
                   end_date = "Oct 19",
                   cohort_website = "https://openscapes.github.io/2023-fred-hutch/")

}
