% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_agenda.R
\name{call_agenda}
\alias{call_agenda}
\title{Create a Call Agenda}
\usage{
call_agenda(
  registry_url,
  cohort_id,
  call_number,
  cohort_sheet = "cohort_metadata",
  call_sheet = "call_metadata",
  website = paste0("https://openscapes.github.io/", cohort_id),
  output_format = md_agenda(),
  output_file = "agenda.md"
)
}
\arguments{
\item{registry_url}{The URL to the cohort registry.}

\item{cohort_id}{The ID of the cohort.}

\item{call_number}{The call number of the agenda.}

\item{cohort_sheet}{The sheet in the registry with cohort information.}

\item{call_sheet}{The sheet in the registry with information about
individual calls.}

\item{website}{A website for the cohort.}

\item{output_format}{The output format of the agenda.}

\item{output_file}{The name of the output file with no file extension.}
}
\description{
Create a Call Agenda
}
