% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ky_create_repo.R
\name{ky_create_repo}
\alias{ky_create_repo}
\title{Create a Git repository for a Cohort}
\usage{
ky_create_repo(name, path = getwd(), edit = TRUE)
}
\arguments{
\item{edit}{}
}
\description{
Create a Git repository for a Cohort
}
