% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kyber-package.R
\docType{package}
\name{kyber-package}
\alias{kyber}
\alias{kyber-package}
\title{kyber: Set Up Learning Cohorts on GitHub}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://openscapes.github.io/kyber/}
}

}
\author{
\strong{Maintainer}: Sean Kross \email{sean@seankross.com} (\href{https://orcid.org/0000-0001-5215-0316}{ORCID})

Authors:
\itemize{
  \item Julia Lowndes (\href{https://orcid.org/0000-0003-1682-3872}{ORCID})
}

}
\keyword{internal}
